% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hurricane.R
\docType{data}
\name{GeomHurricane}
\alias{GeomHurricane}
\title{GeomHurricane}
\format{An object of class \code{GeomHurricane} (inherits from \code{GeomPolygon}, \code{Geom}, \code{ggproto}) of length 4.}
\usage{
GeomHurricane
}
\arguments{
\item{arc_step}{Resolution of the arcs in the wind_radii graph.}

\item{scale_radii}{Implement to scale the radius size.}
}
\description{
ggproto object to display the hurricane radii
plots.
}
\keyword{datasets}
