% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hurricane.R
\name{hurricane_geocode}
\alias{hurricane_geocode}
\title{hurricane_geocode}
\usage{
hurricane_geocode(storm_data, x = "longitude", y = "latitude",
  r = "wind_radius", quadrant = "quadrant", wind_speed = "wind_speed",
  arcRes = 1)
}
\arguments{
\item{storm_data}{Data set from tidy_tracks.}

\item{x}{longitude coordinate of the eye of the hurricanes.}

\item{y}{Latituted coordinate of the eye of the hurricanes.}

\item{r}{Farthest distance for which the wind speed was recorded.}

\item{quadrant}{spatial quadrants on a map (ne, se, sw, and nw) .}

\item{wind_speed}{Wind speed measure.}

\item{arcRes}{Angular degrees'}
}
\value{
A data.frame.
}
\description{
Storm center and wind radius conversion to a set of arc points
}
