% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hurricane.R
\name{read_ext_tracks}
\alias{read_ext_tracks}
\title{read_ext_tracks}
\usage{
read_ext_tracks(file = "ebtrk_atlc_1988_2015.txt", ..., widths = NULL,
  colnames = NULL, degW = TRUE)
}
\arguments{
\item{file}{Filename or raw data}

\item{...}{Alternative arguments to read_fwf()}

\item{widths}{Vector of columns widths}

\item{colnames}{Character vector of column names}

\item{degW}{Degrees West, file coding format for nothern hemisphere.}
}
\value{
A data.frame
}
\description{
Read in ext_tracks data from Colorado State
}
