% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hurricane.R
\name{tidy_tracks}
\alias{tidy_tracks}
\title{tidy_tracks}
\usage{
tidy_tracks(ext_tracks, makeDate = TRUE, keepcols = c("storm_name", "date",
  "longitude", "latitude"))
}
\arguments{
\item{ext_tracks}{Output from the read_ext_tracks function.}

\item{makeDate}{Date columns combination option into a single date column.}

\item{keepcols}{Character vector of the columns  retained from the original data
#'}
}
\value{
A tidy data.frame
}
\description{
Tidy data creation for hurricane data. It takes data from the
read_ext_tracks
}
